Compare usernames without case. Fixes #16
GitHub usernames are not case sensitive and sometimes people will just
type out the name they want (if they're not using the GitHub UI, for
example). But when commenting, the case of the login is always
preserved.

This can lead to situations where a user is mentioned with a different
case than what they use in their username. This catches that appropriately.