Create PR records on synchronize. Fixes #14
This will create run the new PR routines on the synchronize event in the
case that a record for the PR does not yet exist.

This catches the case where a PR fails one of the checks when it is
initially opened and then the author updates the branch, by pushing new
code or force pushing to a different revision. We would previously lose
track of the PR altogether, but now we should keep running the PR
through the machinery regardless.