Repository column for PullRequest and Avoid Commit Author as ReviewRule reviewer
* Add Repository field to PullRequest

Add RepositoryField to PullRequest object to further future support for
multiple repos in a single instance of Cody.

* Avoid Commit Authors as Reviewers for ReviewRules

People who wrote commits should not be the people reviewing the PR if
possible. This PR attempts to select a reviewer for a ReviewRule who is
not current a reviewer and is not a commit author.  It then falls back
to a reviewer that is not a current reviewer. It then falls back to a
random reviewer.