Allow PRs to link review status to other PRs (#62)
Authors can now delegate review of one PR to another. This is mainly for
situations such as backporting a piece of code to historical branches
at the same time as the code is being reviewed as a PR to the main
branch. Insead of requiring reviewers to also review the backport at
the same time as the main PR, authors can delegate the backport's review
status to the main PR. Once the main PR is approved, all delegated PRs
will also be approved and have their status updated.